Some server hax, this is still a mess.
I desperately need to refactor most of this.
We've decided that the server should keep as little game state as
possible; this means that you shouldn't be able to mess with games that
aren't your own, but could build weird mods. Whether that's smart is
something for the ages to decide.